# ARROW-11066: [Java][FlightRPC] fix zero-copy optimization