@emkornfield 2021-02-01
It looks like this might have broken mac os java.   I'll rollback tonight if this is the case and we can open a new pr to fix