@lidavidm 2021-02-01
Yes, I was poking it...sometimes the Debian Java build fails, other times the MacOS build. I'll take a look when I get a chance, though without a MacOS machine, that might take a while.